Bundle the jenkins.io apt repository key within the charm (#20)
Bundle the jenkins.io apt repository key in the charm

The jenkins charm currently fetches the repository key using http at
install time before using it to validate the apt repository. This
leaves it open to man-in-the-middle attack. If an attacker can
redirect jenkins.io to a bogus server, that server could contain
both a signed archive and a public key it was signed with.

This change provides the key within the charm itself so that we're
validating with a known valid key.