Add a jenkins.ready state for use by upper charm layers
I'm building a charm that consumes the jenkins-charm layer, and I want
to be able to set a status message when both the Jenkins layer and my layer
are in ready state. I initially tried to use the individual states already
exposed in the jenkins charm layer along with my own state:

    @when('jenkins.configured.tools',
          'jenkins.configured.admin',
          'jenkins.configured.plugins',
          'mylayer.ready')
    def mylayer_ready():
        status_set('active', 'MyJenkinsLayer is Ready')

However, this could lead to either a message of "Jenkins is running" from
the jenkins layer or "MyJenkinsLayer is Ready" from my layer, depending on
which one reacts last.

This commit adds a new state to the jenkins layer that I can use to guarantee
my reaction runs after the jenkins layer:

    @when('jenkins.ready',
          'mylayer.ready')
    def mylayer_ready():
        status_set('active', 'MyJenkinsLayer is Ready')