Set with_version=False for plugin download.
By default, UpdateCenter will download plugins to a file like
foo-x.y.jpi, which jenkins will unpack to a directory called foo-x.y.
For some versions of jenkins, this causes Jenkins to ignore it in favour
of its built in plugins, which will install.  This change means that
downloaded plugins will be respected by jenkins properly.