Do not restart Jenkins when no change occurs
Due to the way safeRestart behave, any call to the install_plugins
method will result in a restart, effectively putting in queue all
subsequent builds.
If a very long build is running, this prevent any build to run for its
duration, which is impractical.
This is the first part of the mitigation, which will not issue a
safeRestart when there is nothing to change:
* no plugin got updated
* no plugin got removed

A more complex mitigation can be designed later.